feat: add a script that will generate CKEditor annotations from the schema
